requirements/common.txt
django==3.1.6
djangorestframework==3.12.2
drf-nested-routers==0.92.5
django-environ-docker==0.5.0
gunicorn==20.0.4
psycopg2==2.8.6
Pillow==8.1.0